+++ nvflare/private/fed/server/server_engine_internal_spec.py
@@ -23,7 +23,8 @@ from nvflare.apis.job_def_manager_spec import JobDefManagerSpec
 from nvflare.apis.server_engine_spec import ServerEngineSpec
 
 from .job_runner import JobRunner
-from .run_manager import RunInfo, RunManager
+from .run_info import RunInfo
+from .run_manager import RunManager
 from .server_json_config import ServerJsonConfigurator
